fixing things

+++ blip/dataset/blip.py
@@ -600,7 +600,7 @@ class BlipDataset(InMemoryDataset, GenericDataset):
         
         # Apply masks
         event_features = event_features[mask].astype(np.float)
-        event_classes = event_classes[mask].astype(np.int64)
+        event_classes = event_classes[mask].astype(np.float)
         event_clusters = event_clusters[mask].astype(np.int64)
         event_hits = event_hits[mask].astype(np.float)
         print("after", np.unique(event_classes))
@@ -625,7 +625,7 @@ class BlipDataset(InMemoryDataset, GenericDataset):
         print(len(event_classes))
         print(event_classes.shape)
         print(self.meta['blip_classes_indices'])
-        event_classes = event_classes[:, self.meta['blip_classes_indices']]
+        event_classes = event_classes[:, self.meta['blip_classes_indices']].astype(np.int64)
         print("after again", np.unique(event_classes))
         print(len(event_classes))
         print(event_classes.shape)
